requirements.txt
python-dotenv-0.10.3
requests-2.22.0
python-telegram-bot-12.2.0
